requirements.txt
fastapi
uvicorn[standard]
pymupdf
openai
faiss-cpu
tiktoken
python-dotenv
langgraph
langchain-openai
langchain-core
